Update README.md
This is a request to conduct an end-to-end (E2E) test for Instagram's auto-testing feature, specifically for creating a new user account using a temporary email address. The purpose of this test is to ensure that the auto-testing feature is working correctly and efficiently, and that new users can be created seamlessly using temporary email addresses.